Allow additional metadata to define geographic locations
This interprets specific lat/long columns in the CSV/TSV as association a strain with a geographic location.  As this approach defines lat-longs _per-sample_ it is orthogonal to Nextstrain's approach (where we associate coords to a metadata trait). The approach employed here is to create a new (dummy) trait whose values represent the unique lat/longs provided.

If a JSON does not define any lat/longs then the appropriate additional metadata will trigger the map to become available (and displayed).